+++ example/ios/ReactNativeTHEOplayer/AppDelegate.m
@@ -8,8 +8,7 @@
 #import <THEOplayerSDK/THEOplayerSDK-Swift.h>
 #endif
 
-// uncomment for react-native-google-cast
-//#import <GoogleCast/GoogleCast.h>
+#import <GoogleCast/GoogleCast.h>
 
 @implementation AppDelegate
 
@@ -34,12 +33,11 @@
   self.window.rootViewController = rootViewController;
   [self.window makeKeyAndVisible];
   
-// uncomment for react-native-google-cast
-// NSString *receiverAppID = @"CC1AD845"; // default receiver
-//  GCKDiscoveryCriteria *criteria = [[GCKDiscoveryCriteria alloc] initWithApplicationID:receiverAppID];
-//  GCKCastOptions* options = [[GCKCastOptions alloc] initWithDiscoveryCriteria:criteria];
-//  options.startDiscoveryAfterFirstTapOnCastButton = false;
-//  [GCKCastContext setSharedInstanceWithOptions:options];
+  NSString *receiverAppID = @"CC1AD845"; // default receiver
+  GCKDiscoveryCriteria *criteria = [[GCKDiscoveryCriteria alloc] initWithApplicationID:receiverAppID];
+  GCKCastOptions* options = [[GCKCastOptions alloc] initWithDiscoveryCriteria:criteria];
+  options.startDiscoveryAfterFirstTapOnCastButton = false;
+  [GCKCastContext setSharedInstanceWithOptions:options];
   
   return YES;
 }