+++ ios/THEOplayerRCTTrackMetadataAggregator.swift
@@ -24,7 +24,14 @@ let PROP_STARTTIME: String = "startTime"
 let PROP_ENDTIME: String = "endTime"
 let PROP_CUES: String = "cues"
 let PROP_CUE_CONTENT: String = "content"
+let PROP_CUE_CUSTOM_ATTRIBUTES: String = "customAttributes"
 let PROP_SRC: String = "src"
+let PROP_START_DATE: String = "startDate"
+let PROP_END_DATE: String = "endDate"
+let PROP_ATTRIBUTE_CLASS: String = "class"
+let PROP_DURATION: String = "duration"
+let PROP_PLANNED_DURATION: String = "plannedDuration"
+let PROP_END_ON_NEXT: String = "endOnNext"
 
 class THEOplayerRCTTrackMetadataAggregator {
 
@@ -33,7 +40,7 @@ class THEOplayerRCTTrackMetadataAggregator {
         let audioTracks: AudioTrackList = player.audioTracks
         let videoTracks: VideoTrackList = player.videoTracks
         return [
-            EVENT_PROP_TEXT_TRACKS : THEOplayerRCTTrackMetadataAggregator.aggregatedTextTrackListInfo(textTracks: textTracks, metadataTracks: metadataTracksInfo),
+            EVENT_PROP_TEXT_TRACKS : THEOplayerRCTTrackMetadataAggregator.aggregatedTextTrackListInfo(textTracks: textTracks, metadataTracks: metadataTracksInfo, player: player),
             EVENT_PROP_AUDIO_TRACKS : THEOplayerRCTTrackMetadataAggregator.aggregatedAudioTrackListInfo(audioTracks: audioTracks),
             EVENT_PROP_VIDEO_TRACKS : THEOplayerRCTTrackMetadataAggregator.aggregatedVideoTrackListInfo(videoTracks: videoTracks),
             EVENT_PROP_SELECTED_TEXT_TRACK: THEOplayerRCTTrackMetadataAggregator.selectedTextTrack(textTracks: textTracks),
@@ -44,19 +51,19 @@ class THEOplayerRCTTrackMetadataAggregator {
     }
 
     // MARK: TEXTTRACKS
-    class func aggregatedTextTrackListInfo(textTracks: TextTrackList, metadataTracks: [[String:Any]]) -> [[String:Any]] {
+    class func aggregatedTextTrackListInfo(textTracks: TextTrackList, metadataTracks: [[String:Any]], player: THEOplayer) -> [[String:Any]] {
         var trackEntries:[[String:Any]] = metadataTracks
         guard textTracks.count > 0 else {
             return trackEntries
         }
         for i in 0...textTracks.count-1 {
             let textTrack: TextTrack = textTracks.get(i)
-            trackEntries.append(THEOplayerRCTTrackMetadataAggregator.aggregatedTextTrackInfo(textTrack: textTrack))
+            trackEntries.append(THEOplayerRCTTrackMetadataAggregator.aggregatedTextTrackInfo(textTrack: textTrack, player: player))
         }
         return trackEntries
     }
     
-    class func aggregatedTextTrackInfo(textTrack: TextTrack) -> [String:Any] {
+    class func aggregatedTextTrackInfo(textTrack: TextTrack, player: THEOplayer) -> [String:Any] {
         var entry: [String:Any] = [:]
         entry[PROP_ID] = textTrack.id
         entry[PROP_UID] = textTrack.uid
@@ -70,7 +77,7 @@ class THEOplayerRCTTrackMetadataAggregator {
         if !textTrack.cues.isEmpty {
             var cueList: [[String:Any]] = []
             for cue in textTrack.cues {
-                cueList.append(THEOplayerRCTTrackMetadataAggregator.aggregatedTextTrackCueInfo(textTrackCue: cue))
+                cueList.append(THEOplayerRCTTrackMetadataAggregator.aggregatedTextTrackCueInfo(textTrackCue: cue, player: player))
             }
             entry[PROP_CUES] = cueList
         }
@@ -91,16 +98,65 @@ class THEOplayerRCTTrackMetadataAggregator {
     }
     
     // MARK: TEXTTRACK CUES
-    class func aggregatedTextTrackCueInfo(textTrackCue: TextTrackCue) -> [String:Any] {
+    class func aggregatedTextTrackCueInfo(textTrackCue: TextTrackCue, player: THEOplayer) -> [String:Any] {
         var entry: [String:Any] = [:]
         entry[PROP_ID] = textTrackCue.id
         entry[PROP_UID] = textTrackCue.uid
-        entry[PROP_STARTTIME] = THEOplayerRCTTypeUtils.encodeInfNan((textTrackCue.startTime ?? 0) * 1000)
-        entry[PROP_ENDTIME] = THEOplayerRCTTypeUtils.encodeInfNan((textTrackCue.endTime ?? 0) * 1000)
+        var startTime = textTrackCue.startTime
+        var endTime = textTrackCue.endTime
+        if let dateRangeCue = textTrackCue as? DateRangeCue,
+           let programDateTime = player.currentProgramDateTime?.timeIntervalSince1970 {
+            let currentTime = player.currentTime
+            let offset = programDateTime - currentTime
+            startTime = dateRangeCue.startDate.timeIntervalSince1970 - offset
+            endTime = dateRangeCue.endDate != nil ? dateRangeCue.endDate!.timeIntervalSince1970 - offset : .infinity
+        }
+        entry[PROP_STARTTIME] = THEOplayerRCTTypeUtils.encodeInfNan((startTime ?? 0) * 1000)
+        entry[PROP_ENDTIME] = THEOplayerRCTTypeUtils.encodeInfNan((endTime ?? 0) * 1000)
         if let content = textTrackCue.content {
             entry[PROP_CUE_CONTENT] = content
         } else if let contentString = textTrackCue.contentString {
             entry[PROP_CUE_CONTENT] = contentString
+        }
+        if let dateRangeCue = textTrackCue as? DateRangeCue {
+            entry[PROP_START_DATE] = dateRangeCue.startDate.timeIntervalSince1970 * 1000
+            if let endDate = dateRangeCue.endDate {
+                entry[PROP_END_DATE] = endDate.timeIntervalSince1970 * 1000
+            }
+            if let attributeClass = dateRangeCue.attributeClass {
+                entry[PROP_ATTRIBUTE_CLASS] = attributeClass
+            }
+            if let duration = dateRangeCue.duration {
+                entry[PROP_DURATION] = THEOplayerRCTTypeUtils.encodeInfNan(duration * 1000)
+            }
+            if let plannedDuration = dateRangeCue.plannedDuration {
+                entry[PROP_PLANNED_DURATION] = THEOplayerRCTTypeUtils.encodeInfNan(plannedDuration * 1000)
+            }
+            entry[PROP_END_ON_NEXT] = dateRangeCue.endOnNext
+            let customAttributes = dateRangeCue.customAttributes
+            let customAttributesDict = customAttributes.getAttributesAsDictionary()
+            if !customAttributesDict.isEmpty {
+                var attributesEntry: [String:Any] = [:]
+                for (key, _) in customAttributesDict {
+                    do {
+                        // try reading as string
+                        attributesEntry[key] = try customAttributes.getString(for: key)
+                    } catch {
+                        do {
+                            // try reading as double
+                            attributesEntry[key] = try customAttributes.getDouble(for: key)
+                        } catch {
+                            do {
+                                // try reading as data
+                                attributesEntry[key] = try customAttributes.getBytes(for: key)
+                            } catch {
+                                print("Unable to extract customAttribute from DateRange cue. Content is limited to String, Double or Data.")
+                            }
+                        }
+                    }
+                }
+                entry[PROP_CUE_CUSTOM_ATTRIBUTES] = attributesEntry
+            }
         }
         return entry
     }